Bifurcated the logic into rest and service layer.
